Cached html is valid `ASCII` so a `[u8]` is being used back again instead of `*const str`. `(usize, usize)` tuples replaced with u32 since max size of cache is 6MB, no index can be greater than `6,000,000`, which fits in `u32` and takes half of a `usize` space.

+++ tutorlolv2_frontend/src/components/calculator/damage_stack.rs
@@ -12,7 +12,7 @@ use yew::{
 
 fn base_content(
     img_path: ImageType,
-    offsets: Option<&'static (usize, usize)>,
+    offsets: Option<&'static (u32, u32)>,
     onclick: Callback<MouseEvent>,
     content: Option<Html>,
 ) -> Html {
